fix(opentelemetry): move diag log configuration to setup utility
